pyproject.toml
[project]
name = "notes-agent"
version = "0.1.0"
description = "Nwokike/Notes-Agent: Autonomous AI agent for writing community notes on Igbo archives. Developed by Kiri Research Labs."
authors = [{ name = "Onyeka Nwokike" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.12.0",
    "fastapi>=0.135.2",
    "google-adk>=1.28.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.8.0",
    "litellm>=1.82.6",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]
